refactor: restructure into FSD layers (app/features/shared) with slice public APIs
