+++ src/features/activity/index.ts
@@ -1,0 +1,2 @@
+export { ActivityService } from "./activity.service";
+export { MetadataUpdateService } from "./metadata-update.service";
