+++ src/features/clock/clock.constants.ts
@@ -1,0 +1,3 @@
+export const STATUS_BAR = {
+	placeholder: ":)",
+} as const;
