+++ src/features/clock/index.ts
@@ -1,0 +1,1 @@
+export { ClockStatusService } from "./clock-status.service";
