+++ src/features/most-edited/index.ts
@@ -1,0 +1,3 @@
+export { MostEditedService } from "./most-edited.service";
+export { MostEditedView } from "./most-edited.view";
+export { VIEW_TYPES } from "./most-edited.constants";
