+++ src/features/most-edited/most-edited.service.ts
@@ -1,7 +1,7 @@
 import { App, TFile } from "obsidian";
-import { FRONTMATTER_FIELDS, MOST_EDITED_VIEW } from "../../constants/plugin.constants";
-import { getNestedFrontmatterValue } from "../../utils/frontmatter-path";
-import { isFileIgnored } from "../../utils/ignore-rules";
+import { getNestedFrontmatterValue } from "../../shared/lib/frontmatter";
+import { isFileIgnored } from "../../shared/lib/ignore";
+import { FRONTMATTER_FIELDS, MOST_EDITED_VIEW } from "./most-edited.constants";
 
 export interface MostEditedEntry {
 	file: TFile;
