+++ src/features/settings-tab/index.ts
@@ -1,0 +1,1 @@
+export { TimeThingsSettingsTab } from "./settings.tab";
