+++ src/features/settings-tab/settings-tab.constants.ts
@@ -1,0 +1,6 @@
+export const SETTINGS_LINKS = {
+	momentFormatDocs:
+		"https://momentjscom.readthedocs.io/en/latest/moment/04-displaying/01-format/",
+	nonTypingEditingDocs:
+		"https://github.com/DynamicPlayerSector/timethings/wiki/Calculating-your-non%E2%80%90typing-editing-percentage",
+} as const;
