+++ src/shared/config/index.ts
@@ -1,0 +1,2 @@
+export { DEFAULT_SETTINGS, normalizeUpdateInterval } from "./settings.defaults";
+export type { TimeThingsSettings, TimeThingsSettingsManager } from "./settings.types";
