+++ src/shared/config/settings.types.ts
