+++ src/shared/lib/frontmatter/index.ts
@@ -1,0 +1,9 @@
+export {
+	findFrontmatterEndLine,
+	findFrontmatterFieldLine,
+	hasFrontmatter,
+	readFrontmatterFieldValueAtLine,
+	setFrontmatterFieldValue,
+} from "./editor-frontmatter";
+export { getNestedFrontmatterValue, setNestedFrontmatterValue } from "./frontmatter-path";
+export type { FrontmatterObject } from "./frontmatter-path";
