+++ src/shared/lib/ignore/ignore-rules.ts
