+++ src/shared/lib/ignore/index.ts
@@ -1,0 +1,6 @@
+export {
+	formatIgnoreList,
+	isFileIgnored,
+	normalizeIgnorePath,
+	parseIgnoreList,
+} from "./ignore-rules";
