+++ src/shared/ui/suggesters/index.ts
@@ -1,0 +1,2 @@
+export { FileInputSuggest } from "./file-input-suggest";
+export { FolderInputSuggest } from "./folder-input-suggest";
